LFD v1.2.1
Live Footage Detection has a slight improvement, where now the user will be detected with a blue box and the object with a cyan coloured box. Only when a user and atleast one object is detected, the 'Capture button' enables and when clicked will identify and let the user know that the user has been identified and so is the object (${user) and ${item}). Have changed the object detection model to google's Mediapipe object detection model for lightweight web based model.